Updated openstack/openstack Project: openstack/tempest  08f5dd54ed18c5b84cae2e673e9681adcdd35176
Stop swift resource leaking even if an error occurs

In some test, if an error occurrs in setUpClass(), some resources
are not deleted because of tearDownClass() is not called.
This commit fixes it but doesn't touch the code of creating only one
resource without waiting for the resource creation.

 tempest/api
  +---- baremetal (unnecessary)
  +---- compute
  +---- data_processing (unnecessary)
  +---- database (unnecessary)
  +---- identity
  +---- image (done)
  +---- network
  +---- object_storage         <-- this patch
  +---- orchestration
  +---- queuing (unnecessary)
  +---- telemetry (unnecessary)
  +---- volume (done)

Partially Implements: blueprint stop-leaking

Change-Id: I69dd5bf9dbfa04e463b63f4f236fc0aa9e70e79a