Updated openstack/openstack Project: openstack/requirements  8afc9b425ae2d79334352cdad89ee5ab8e3e2d82
Bump python-keystoneclient to >= 0.8.0

Starting from python-keystoneclient 0.8.0, the preferred way to
specify authN and identity endpoints is by using auth_uri and
identity_uri settings.  Otherwise, warnings are issued. To be able to
migrate to brand new world, we need to bump to the new version where
these settings were introduced.

Change-Id: Icbb9c3af9110e405d8b44bda529ccf1b3c1ae3a0