Updated openstack/openstack Project: openstack/tempest  89d4ec73776dcd72f870d4283c6510e542cfabff
Add unit tests for _get_unversioned_endpoint

This commit adds the first unit tests for the verify_tempest_config
script which covers the code used to format an unversioned endpoint
from a versioned one returned from the keystone catalog. This also
breaks that section of the script off into a separate method so that
mocks were not needed for such a simple test case.

Partially implements bp: config-verification

Change-Id: I6f48034af5fc187f6ae7d5c9b3dccb61f309ae25