Updated openstack/openstack Project: openstack/nova  0e479b516c8595c54a0bfe95b7a327eff0bbf7d6
xenapi: fixup agent tests

The root cause of the slow agent tests was not mocking this method:
compute_utils.add_instance_fault_from_exc

This cleans up the tests to share a bit more code, and ensures they
actually check that the appropriate exception is raised.

This change adds a log message to report errors that could occur in
running systems that also helped identify this issue in the unit tests.

test_spawn_fails_agent_not_implemented was a duplicate test and has been
removed.

Change-Id: I6fd31520fffd5e888ce0c2b58af2715f0d929ce1
Closes-Bug: #1279317