Updated openstack/openstack Project: openstack/python-keystoneclient  a03c6b9f346dfd6b838fb8175bc50191642576a2
auth_token configurable check of revocations for cached

The auth_token middleware would fail if it couldn't fetch the
revocation list. If the system is configured for UUID tokens then
the revocation list may not be available.

With this fix, the revocation list will only be checked for
cached tokens if the new check_revocations_for_cached option
is set to True.

Also, this change prevents the revocation list from being
checked twice for a PKI token that's validate off-line.

Change-Id: I5408bbe12aefda608ebcb81cf3c7ef068b2bf2f6
Closes-Bug: #1312858