Updated openstack/openstack Project: openstack/keystone  2c85300381714257254e7cb58602c2f182955b39
Allow 'description' in V3 Regions to be optional

Currently, description is required for POST requests to regions when
using the SQL backend. According to the Identity V3 API spec,
description is not required. If a user creates a region without
specifying a description for the region in the request, we can make sure
it is sent to the database as an empty string.

Change-Id: Ia6fb1a2eee6b471c0c44a66590f53d061bcd8643
Related-Bug: 1284972