Updated openstack/openstack Project: openstack/heat  25365338cb9577524b361d6d5f560a6d6d717c5e
Fixes stack rollback failure with AWS::EC2::Instance

This patch would fix the stack rollback failure that
happens with an errored instance in a stack.
When the instance creation fails, vm status/sate would
be 'ERROR'. After delete is issued to nova, it would take
little time to change from 'ERROR' to 'DELETED'. By
checking for 'ERROR' status and throwing an exception,
it takes the create 'ERROR' status and fails to rollback
the stack.

Change-Id: Ida81117504398e0406d7b6d1e0f277af2acb0081
Closes-Bug: #1312636