Updated openstack/openstack Project: openstack/keystone  bab63bff9dc4fb94912f1e9b8a7bba8445f34fd5
Remove the configure portion of extension docs

The docs themselves have little to do with configuration,
and mostly describe how to enable the extension.
Suggest removing the configuration part of the file rename.
Also motivating this change, is the fact that we are including
a doc that helps to configure the federation extension
(https://review.openstack.org/#/c/89220/).

Change-Id: I0fe845ac9bdfede330a7eb56fee6f971290789b3