Updated openstack/openstack Project: openstack/horizon  6f212f85874fc86cfff20847fe89fe3b5a636503
Avoid attr error on database backup page

For some reason (and I genuinely do not have a clue why) some
database-backups don't appear to have an instance attribute. This causes
a problem when loading the table data, as the db_name helper-function
raises an AttributeError.

It appears that we're prepared for obj.instance.name to not exist, but
if obj.instance doesn't exist, then we raise the attribute error.

The solution is to return the obj.instance_id in both cases.

Change-Id: I77a30e4aeb97744aa14151eca0b4c011476fb08d
Closes-Bug: 1312235