Updated openstack/openstack Project: openstack/tempest  4c578bae28d074b6c15ff69f53193f94d70a3d5c
Add a common class for Nova v2/v3 API tests

Now there is a lot of copy&paste test code for Nova v2/v3 API tests.
In addition, we need to add many checks for API test coverage.
As the result, we should apply the same changes to v2 and v3 tests.

For improving the maintenance, this patch changes a base class to
share it between v2 and v3 tests. The existing BaseV2ComputeTest and
BaseV3ComputeTest inheritate from it now, and we will replace each
API test with a base class directly. As a sample, this patch applies
this class to test_availability_zone of v2 and v3.

This patch changes the test directory structure like:
 * tempest/api/compute/ : common test files like this test class
 * tempest/api/compute/v2/ : v2 API specific test files
 * tempest/api/compute/v3/ : v3 API specific test files

After applying this class to all Nova API tests, we will be able to
remove current BaseV2ComputeTest and BaseV3ComputeTest classes.

Partially implements blueprint nova-api-test-inheritance

Change-Id: I4f9c3f58c37fc459976c2e7dc36bfc90f5add3f5