Updated openstack/openstack Project: openstack-dev/devstack  1bb05a96316e3bd4069d30e91a3588340aa71edf
Fix logging for keystone under mod_wsgi

When Keystone is configured under mod_wsgi, Keystone configuration 
requires debugging to be enabled as the debug flag cannot be passed
via cli arguments.

Because Apache 2.2 (default for a number of distributions) does not 
support passing an explicit format to the ErrorLog, Keystone under 
mod_wsgi ends up with duplicated timestamps in the Keystone log. 
This patchset eliminates the %(asctime) and %(msecs) substitution 
targets from the log format options when Keystone is configured 
under mod_wsgi.

Change-Id: I0bf32db6e53cb653921162cb1957eb988b819815