Updated openstack/openstack Project: openstack/heat  aca822a7a127ef880740ac3d206408f5e71ec782
Deprecate Stack.resolve_runtime_data()

This no longer needs to be done in the context of a stack, because all of
the required context is embedded in the lazily-evaluated functions at the
time they are parsed into Functions. Reduce confusion by eliminating all
uses of it and marking it as deprecated.

Change-Id: I9d89b5201df5b05d03a295eaf6605da65397f565