Updated openstack/openstack Project: openstack/nova  dcd024ed7623a54c0d9c6943a57644b7b448666d
Dedup devref on unit tests

We already have a document on how to run unit tests, so
development.environment should just point to that document instead of
trying to summarize it.  Keeping this information in one less places
makes it easier to make sure the page is up to date (which
development.environment is not -- no mention of tox).

Change-Id: I9aecf439b6145b035b3457559ddb187058f5057b