Updated openstack/openstack Project: openstack/tempest  f7289da165b9f506394b6bb974b10b6c3225ba71
Add compute service tags to ListSnapshotImagesTest

The tests in ListSnapshotImagesTest all require nova because they
create a server in setUpClass and then snapshot it so they can test
filtered image lists. This commit adds the appropriate service tag so
the tests are identified as using nova.

Partially implements bp add-service-tags

Change-Id: Idec3110e8c1edfc34b77bf08bb927a8c10596914