Updated openstack/openstack Project: openstack/heat  861bfb215ef25fa62a79c3f4d0792499b34d4464
Revert "Support ... x-forwarded-for header"

This reverts commit e3679217d1d4b1821f4fa0bd9894b761edfe4ba9.

This commit broke using the heat API when client requests were
behind a proxy (this sets the X-Forwarded-For header)
and effectively causes the heat-api to return the clients
IP address back in the request via the location header.

I believe the original (unapproved) heat blueprint for
this feature should have used X-Forwarded-Host instead.

Closes-bug: #1316754

Change-Id: I1b38ececf59545a3c13c7ebf2c33be97f4332ac0