Updated openstack/openstack Project: openstack/nova  a63a7ecd5536753c59ac0ee2e50d63619ac7fcd6
Remove and block DB access in dhcpbridge

This fixes an issue of direct DB access from happening in dhcpbridge
by using an object method instead. For this to work and to prohibit
future DB access from happening, the objects indirection API is set
and DB access is blocked in a similar way to how it is done in
nova-compute and nova-network.

Change-Id: I5e3152aad13343a6a1ecd9b5b511429159c9d4c2
Closes-Bug: #1316074