Updated openstack/openstack Project: openstack/cinder  b7e126e17a34a287dc1ed324ddde22c5806b5421
Implement validate_connector in FibreChannelDriver

The base FibreChannelDriver didn't implement the
validate_connector() method. It fell back to the
parent class which simply does a pass.  Now the
FCDriver checks to ensure that the connector has
wwnns and wwpns and raises an exception if either
one is empty or not set.

Closes-Bug: #1294812

Change-Id: I1e63c4718d24bdb3ead78de871971cdc498f6b01