Updated openstack/openstack Project: openstack/horizon  28ae28b59f1ea740e0e2ed6c1e33c5e7a545467a
Allow static files to be exposed for any plugin

Change the parsing of plugin configuration files to allow
a new Django application to be added. This enables the static file
finder to find the plugin's static files when they are referenced.
Previously only plugins that defined a new dashboard could do this.

Change-Id: I69bd8f24fa0855174a05203eabe7c932246dda21
Closes-Bug: #1309183