Updated openstack/openstack Project: openstack/nova  ac37b0784c2a180ed1e70db0116cbf55a39f17a5
Make vmwareapi/vmops use Instance object for updates

This converts the one use of virtapi.instance_update() in vmops to
use the Instance object instead of the old conductor method.
Unfortunately, the affected method was apparently not tested at all,
so this adds a simple test for it.

Related to blueprint virt-objects-juno

Change-Id: I21ed4903a0c8d98c8776d358d8273077cd979fdf