Updated openstack/openstack Project: openstack/horizon  4cff4e067625b3567fd45f64494a35cfc306a60e
Use floats instead of integer in memory calculation of admin dashboard

In the case that hypervisors have large amounts of memory, you'll find that
the dashboard reports improper memory usage when in the TB and PB range.
For example, 1.95 TB as an integer equals 1 so we convert to float instead.

Fixes bug: 1313009

Change-Id: Ie107986fa3ed9aad5c68c34353cdfc5943788069