Updated openstack/openstack Project: openstack/tempest  bcefa3d8d55d5d0bac4b35bf5e1493c68adad85a
Add a common admin class for Nova v2/v3 API tests

For improving the maintenance, this patch changes a common admin class
to share it between v2 and v3 tests. In addition, this patch applies
this class to test_availability_zone of v2 and v3 as a sample.

After applying this class to all Nova API tests, we will be able
to remove current BaseV2ComputeAdminTest and BaseV3ComputeAdminTest
classes.

Partially implements blueprint nova-api-test-inheritance

Change-Id: Iede054d45c18922102cdf17938fde0e33f1df16c