Updated openstack/openstack Project: openstack/heat  b84816fee37df20d8754905ef6055371690f92f2
Make a top level contrib doc header

This is to just make it clear that there are other resources
besides Rackspace ones.

This is not really a bug as both the rackspace and docker
headings are the same level, but when going to the page it
is easy to see why the bug was made.

Change-Id: I76a8bcf63b7c976c4fcf0baecfd58e5fcbb65d96
Fixes-bug: #1311565