Updated openstack/openstack Project: openstack-dev/devstack  8cedabcea8bb446f1c29aab42fbcbf5a87218f7f
stack.sh fails for Cisco plugin with missing cisco_plugins.ini error

When the Cisco Nexus monolithic plugin is configured in the localrc,
stack.sh fails with the error message:
     sed: can't read /etc/neutron/plugins/cisco/cisco_plugins.ini:
          No such file or directory

This failure was introduced with the merge of change set 76992
(bug #1285884). Change set 76992 removed the copying of
Q_PLUGIN_EXTRA_CONF_FILES from the neutron project area to directories
under /etc/neutron. This copy from neutron to /etc is required by the
Cisco Nexus monolithic plugin, since this plugin follows the original
model for generating extra config files based on localrc settings:
 - Copy default config file(s) from neutron project to /etc using
   a relative path (relative to neutron and relative to /)
 - Modify the copies of the config file(s) using localrc settings
 - Add a --config-file command line setting for neutron server which
   points to the modified config file(s) under /etc

The fix is to add the copying of the default versions of the extra
config files from neutron to /etc into the cisco devstack script.

Change-Id: Ifbae197ed591f7cccfd6d51eae2b21e31eb0a409
Closes-Bug: #1295411