Updated openstack/openstack Project: openstack/horizon  ecd49db80117b42bb56ab2e4350c1f5043ae2730
ValueError when cookie pagesize is empty string

If a user somehow gets the value of 'horizon_pagesize' set to "", the
method horizon.utils.functions.get_page_size will raise a ValueError
because it can't convert the empty string to an int. Clearly this would
also happen if the value gets set to any other non-valid int-literal.

Change-Id: I4b1dd07df75d931f8c61823c93c5dd6ecc145686
Closes-Bug: 1316331