Updated openstack/openstack Project: openstack/heat  61b0860ddceabdf240fd20954ce3c4d74f9a52a2
Make Server metadata a passive store again

This change reverts the software deployments change which polls
the deployments API whenever a Server resource metadata is read.

The Server metadata has reverted to being a standard resource
metadata store, and the deployments are pushed into this metadata
directly in the database whenever the deployments API is used to
create or update a deployment.

As part of the effort to not require a Stack.load for polling metadata
this change is for
Partial-Bug: #1306743

Change-Id: Ib461e1bab816d4b9e332786004c53a2f73d773e8