Updated openstack/openstack Project: openstack/trove  aa481793cbebcfe9835cde2bc599ecb1f9e9a00d
Migrate v20 not handling exception

Version 20 upgrade creates 2 tables but the downgrade doesn't
remove them (for a stated reason). If you downgrade and then
upgrade, the upgrade will fail because the tables already exist.
This fix add a try block around the upgrade to log an info
message instead of failing.

Change-Id: I37fbe002da7dc764d229575149e4062e38f75881