Updated openstack/openstack Project: openstack-dev/devstack  f2fd04924f26aef4f74fa1b192e1fa1d76e1546c
Add an option in lib/swift to increase max_header_size

With the increasing number of entries in the service_catalog header,
some projects will have a header bigger than 8192 and Swift cannot
handle headers of this size. Solum is the first one to have this issue

As a temporary hack to fix Solum's gate, we are increasing
the value of the max_header_size parameter in our "lib" script.

This patch makes this settings global to all projects using devstack,
so that in a near future, if a project has a service_catalog header
bigger than 8192, it doesn't have the same problems we had.

Closes bug 1316585

Change-Id: I5c15f8f9531839353c3d99e46d920688058ec33c