Updated openstack/openstack Project: openstack/cinder  1eda138be81b2405969b80c00f30ba237e250fcd
BrcdFCSanLookupService should allow customize host key and policy

In BrcdFCSanLookupService, the initialization should allow the
customization of the known_hosts_file and missing_key_policy so that the
hosts key and missing policy can be customized according to the
different scenario and customer aspect. This will not change the default
behavior when no argument is given, but more flexible to allow the
caller to give more options according to different requirements.

Closes-Bug: #1320050
Change-Id: If5767f63ccd2cde5fbea30a6154acf4d28f662b6